requirements.txt
streamlit
Pillow
opencv-python
numpy
matplotlib
plotly
streamlit-image-coordinates
